+++ src/main/java/com/unibuc/management/controllers/DoctorController.java
@@ -5,9 +5,14 @@ import com.unibuc.management.services.DoctorService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
+import org.springframework.security.access.prepost.PreAuthorize;
+import org.springframework.security.core.Authentication;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
+import java.util.Optional;
+
 @RestController
 @RequestMapping("/api/doctors")
 public class DoctorController {
@@ -19,7 +24,9 @@ public class DoctorController {
     public ResponseEntity<List<Doctor>> getAllDoctors() {
         return new ResponseEntity<>(doctorService.getAllDoctors(), HttpStatus.OK);
     }
+
     @PostMapping
+    @PreAuthorize("hasRole('DOCTOR')")
     public ResponseEntity<Doctor> createDoctor(@RequestBody Doctor doctor) {
         Doctor createdDoctor = doctorService.saveDoctor(doctor);
         return new ResponseEntity<>(createdDoctor, HttpStatus.CREATED);
@@ -33,8 +40,18 @@ public class DoctorController {
         }
         return new ResponseEntity<>(doctor, HttpStatus.OK);
     }
+
     @PutMapping("/{id}")
+    @PreAuthorize("hasRole('DOCTOR')")
     public ResponseEntity<Doctor> updateDoctor(@PathVariable Integer id, @RequestBody Doctor doctorDetails) {
+        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+        String currentUsername = authentication.getName();
+        Optional<Doctor> currentDoctorOpt = doctorService.getDoctorByUsername(currentUsername);
+
+        if (currentDoctorOpt.isEmpty() || !currentDoctorOpt.get().getId().equals(id)) {
+            return ResponseEntity.status(HttpStatus.FORBIDDEN).build();
+        }
+
         Doctor updatedDoctor = doctorService.updateDoctor(id, doctorDetails);
         if (updatedDoctor == null) {
             return new ResponseEntity<>(HttpStatus.NOT_FOUND);
@@ -43,7 +60,16 @@ public class DoctorController {
     }
 
     @DeleteMapping("/{id}")
+    @PreAuthorize("hasRole('DOCTOR')")
     public ResponseEntity<Void> deleteDoctor(@PathVariable Integer id) {
+        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+        String currentUsername = authentication.getName();
+        Optional<Doctor> currentDoctorOpt = doctorService.getDoctorByUsername(currentUsername);
+
+        if (currentDoctorOpt.isEmpty() || !currentDoctorOpt.get().getId().equals(id)) {
+            return ResponseEntity.status(HttpStatus.FORBIDDEN).build();
+        }
+
         boolean isDeleted = doctorService.deleteDoctor(id);
         if (!isDeleted) {
             return new ResponseEntity<>(HttpStatus.NOT_FOUND);
